+++ category/interfaces/errors.go
@@ -1,0 +1,10 @@
+package interfaces
+
+import (
+	"errors"
+)
+
+var (
+	ErrCategoryNotFound = errors.New("category_not_found")
+	ErrCategoryGeneral  = errors.New("category_general_error")
+)